Rework daemon life cycle management
Let the daemon always listen on both 'name-appeared' and
'name-vanished' and enable opening new sessions when all required D-Bus
services become available, and close sessions when they disappear.

Add a 'terminate' action so that the daemon can be told to terminate.
This needs a centralized place listing open sessions, and a way to let
the daemon to close sessions on shutdown. gnome-remote-desktop-control
--terminate can be used to tell the daemon to shut down.