vnc: Add 'prompt' and 'password' authorization methods
The 'prompt' authorization method will prompt the user (using libnotify)
asking them to either accept or deny a request to view and control the
session.

The 'password' authorization method will instead ask the remote user to
provide the set password. The password is specific to VNC and is stored
in the keyring and accessed using libsecret.