project: Compile gschema on install
When a gschema is set, it must be compiled at install
time for DConf be able to recognize it. The current
install procedure doesn't do that, which causes an
immediate segfault when starting the remote desktop
service.

Fix that by adding a post-install script that compiles
the schemas after installing them.