vnc: Close session in idle callback when client disappears
To avoid re-entry in libpipewire when a client disappears during a
render call, handle VNC clients being gone in a idle callback so that we
don't ever destroy a stream while processing a PipeWire buffer.

This also means we need to handle the socket errors more gracefully, as
we'll see those on the socket handler after the client was gone but
before we stopped the session.