vnc: Postpone resizing framebuffer until encodings known
The client will send a list of extensions, including one that tells the
server that the client supports resizing the framebuffer. If we resize
the framebuffer before the server has received this list, it'll drop the
resize on the floor.

To not have the resize request dropped on the floor, wait until the
client has sent the extensions list before resizing the framebuffer.
This is done by checking the 'preferredEncoding' variable of the
rfbClientRec, which is set as a side effect of receiving the protocol
message that also contains available extensions.