session-vnc: Set our own password handling function up front
libvncserver decides whether to register a auth security handler
depending on whether the password data is set or not. When we use the
prompt auth method, we don't want to ask for password, so set the
password data to NULL.

Also, to be a bit more in control of the password mechanism, always set
the password function up front, instead of just when the client uses the
password prompt.