vnc/pipewire-stream: Drop NULL buffer
If we skipped a buffer due to a newer one being available, we'll still
get a 'new buffer' event, but as we've already consumed it, there will
be no new buffer available to process. Handle this by just ignoring
returning early if we end up with no new buffer available.