vnc/pipewire-stream: Copy buffer content in real-time thread
This changes the stream behavior to copy the buffer content immediately
as it is received, making it less likely that mutter will have had time
to over-write the contents with a new frame. It doesn't cause any extra
copying, but does result in a full frame allocation each frame, instead
of just a copy.