settings: Add getter and setter for RDP settings
Add getter for the RDP port, server certificate and server keyfile.
Also add a setter in order to be able to update the RDP server port
in the settings.
The path for the server certificate and the server keyfile are both
stored via gsettings.

Also add a SecretSchema in order to be able to get the username and
password for RDP.