rdp: Add initial implementation for RDP
Similar to the VNC implementation, the RDP implementation gets a
RDP server, which will listen for new connections and create a new
RDP session for the new client, a RDP session which will handle
anything related to the client (such as sending a new Frame, taking
care of the authentication of the client or in the future also
handling any keyboard or mouse input) and a PipeWire stream,
which gets the frames, that the the client should display to the
user.

In order to be able to communicate with the client, the library
FreeRDP is being used.
The first thing that will happen when a new RDP session is being
created is passing the server certificate to the client for TLS
Security and making sure that the client meets all requirements.
These requirements are a valid authentication, which is done via
NlaSecurity in FreeRDP (valid username and password), and the support
of specific codecs in order to able to the view the content sent
by the server.
Currently the server requires the client to support either the
NSCodec or the RemoteFXCodec.
Once the requirements are fulfilled, the server starts sending
the frames to the client which will then display them.
If the requirements are not met, the server closes the connection
to the client.
Currently, the server only supports to view only the content.
Any remote input, such as keyboard or mouse input, is being discarded
by the server.