session-rdp: Add implementation for extended mouse events
Similar to the implementation of normal mouse events, forward extended
mouse events to mutter via the RemoteDesktop interface.
Extended mouse events are button events of the BTN_SIDE- and the
BTN_EXTRA-button.
Usually, these buttons are used as Back- and Forward-button when using
(file-)browsers.
Also similar to the HasHorizontalWheel property, advertise the
HasExtendedMouseEvent property to the client in order to be able to
receive these events.