session-rdp: Add API to move the mouse pointer
Add an API to move the mouse pointer on the window of the connected
client.
It is however not defined how the client handles the pointer position
update since the implementation of the pointer position update lies in
the hands of the client.
The client could also simply discard the data and do nothing.