pipewire: Add pipewire-utils
When we implement other backends other than the already existing VNC
backend, some functions and attributes in the pipewire backend of the
new backend are either simply the same or can (and should) be shared
with the other pipewire implementation due to unnecessary duplication.
Pipewire-utils picks these functions and attributes up.