rdp: Add handling for RDP SAM files
FreeRDP needs for NlaSecurity a SAM file which contains the accepted
usernames and their associated passwords in a NTLM hash form.
Unfortunately, this file cannot be passed via a file descriptor to
FreeRDP.
Instead, we have to create a temporary file in RAM with the required
data, pass it to FreeRDP and delete it afterwards.