Add damage-utils
Encoding and sending frames can take a lot of time and sometimes the
content that is sent to the client might also be larger than it is
intended.
Using damage regions can both reduce the encoding time of a frame and
the amount of data that will be transmitted to the client by only
updating the regions that actually have changed.

damage-utils provides functions that create these regions for two
provided frames.