rdp-pipewire-stream: Update mouse pointer and mouse pointer position
Previously, the RDP backend did not send the new pointer bitmap and
pointer position to the client which led that the client would always
use their system pointer instead of our pointer bitmap.
The RDP backend however, added recently an API to send these values to
the client.

So, update these values using that API.