session-rdp: Fix inconsistency regarding length to allocate
session-rdp has the convention that the count of something that will be
allocated means the amount of something that will be allocated without
including the resulting size itself, while using length when allocating
something means the size of the data in the memory.
While using `sizeof (uint8_t)` doesn't change the current behaviour, it
still introduces this inconsistency.

Fix this inconsistency by moving the `sizeof ()` part into the length
variable.