Add clipboard class for RDP
Add the implementation of grd-clipboard-rdp, a class that is derived
from grd-clipboard, which implements the RDP specific clipboard parts.

The communication for the clipboard in RDP consists out of several
message types.
Six out of them are necessary when the clipboard has been established:

1. Format List
2. Format List Response
3. Format Data Request
4. Format Data Response
5. File Contents Request
6. File Contents Response

The Format List tells the other peer what kind of data (i.e. the data
format) lies on its clipboard.
It is sent when the clipboard of a peer changes or as part of the
initialization sequence in order to synchronize the initial state of the
clipboard.

The Format List Response is the answer for the Format List and only has
a meaning when the CB_RESPONSE_FAIL flag is set (rare case).
In such case g-r-d will answer all Format Data Requests and File
Contents Requests with the fail flag (CB_RESPONSE_FAIL) set, as the
protocol requires it (3.1.5.2.4 Processing a Format List Response PDU).
This state is reset when g-s advertises a new Format List.
All types of Response messages either contain the flag CB_RESPONSE_OK
(indicating success) or the flag CB_RESPONSE_FAIL (indicating failure).

The purpose of the Format Data Request is to request the data of a
specific (previously announced) format type.
The Format Data Response can then either contain the CB_RESPONSE_OK
flag in addition to the requested data, or the CB_RESPONSE_FAIL flag.

The File Contents Request and Response are similar to the Format Data
ones, with the difference that the File Contents Request and Response
handle file data and are dependant on a previous Format Data Response
that was sent as an answer to Format Data Request that requested the
file list.

All requests by either g-s or the remote desktop client are handled in
the main thread as dbus handling (even _sync calls) is still handled in
the main thread too, and calling these _sync calls from other threads
can potentially lead g-r-d into a deadlock.
This means that requests and responses from a remote desktop client
push a task to the main thread to ensure that no deadlock happens.

Additionally, use events (provided by WinPR) to control the state of
such requests and responses to ensure that all messages are handled in
their correct order and to ensure that specific actions (like
requesting some clipboard content from the clients) are only handled
when they are allowed to do so (no pending Format List).