mime-type: Also save external clipboard format for RDP clipboard
RDP supports multiple text types from which we can convert our
clipboard format from.
However, these formats are not the same as the ones that are supported
on Linux.
In order to support these formats nevertheless, save the external id
for the clipboard format additionally to the internal mime type, in
order to be able to identify the external format correctly when
converting the external format into a format that is supported on
Linux.