Add clipboard class for VNC
Add the implementation of grd-clipboard-vnc, a class that is derived
from grd-clipboard, which implements the VNC specific clipboard parts.

In contrary to the RDP clipboard, VNC does not handle mime-types other
than plain text.
In addition to that, the content is directly transferred to the other
peer as soon as the clipboard changes.