rdp-sam: Add space for the null terminator in filename
strlen () does not count the null terminator as part of the actual
string and therefore the filename of the SAM file currently does not
have space for the null terminator allocated.
While this didn't have caused a crash so far, fix this still by adding
space for the null terminator.