rdp-pipewire-stream: Correct memcpy size calculation
The previous commit addressed an issue with memcpy size
calculation in the vnc backend.  The rdp backend inherited,
the same bug: a potential crash when the source row stride
is less than the destination row stride.

This commit introduces the same fix, using the width of the
row pixel data, only, when performing the copy.

https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/18