vnc-pipewire-stream: Correct memcpy size calculation
commit 78c5bcb181fe2b0b9fc17eea696feac8b504df54 introduced code
to try to account for a row stride mismatch between source and
destination.

That commit, however, erroneously copies destination stride bytes
from the source, which will lead to out of bounds access if the
source has a smaller row stride.

This commit addresses the problem, by instead, copying just the
pixel data, and none of the row margin.

https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/18