session-rdp: Run socket handling in a separate thread
Currently, everything in gnome-remote-desktop is handled inside of one
thread with the exception of the worker threads when encoding a frame
for a RDP client and the PipeWire thread, where we receive the
frame buffers.
For clipboard support in RDP, we have to ensure that only one data
request happens at the same time, since normal clipboard content
(everything except files) is not bound to some identification number
to assign a response to a request.
This implies that data requests cannot be (easily) implemented in an
asynchronous way and that we have to wait for the response in the same
thread.
However, this will also block any incoming data that will be handled in
`handle_socket_data()`, which will unavoidably lead to a deadlock.
Similar situations will happen later with the graphics pipeline or
other virtual channels in RDP.

Moreover, the current socket handling via `handle_socket_data()` causes
other errors:
When a new remote desktop client connects, `handle_socket_data()` is
not always called, when it needs to be called.
This leads to a problem, where the remote desktop client waits for its
activation procedure to complete, but glib cannot realize that it has
to call `handle_socket_data` to complete the procedure.

To solve these situations, run the socket handling in a separate thread
to not be able to run into a deadlock, and handle the socket handling
with functions provided by FreeRDP and WinPR.
The socket thread is like `handle_socket_data()` completely event-
based.
For the socket handling, FreeRDP allows us to listen on its transport
event handles.
With these event handles, g-r-d can now properly handle the activation
process for clients at any time (which is not always possible with the
socket handling functions provided by glib).
Due to the handling of input events now happens in the socket thread,
use the previous implemented RdpEventQueue to queue all input events
on the main thread, since dbus handling still happens in the main
thread.

In addition to that, this also simplifies some areas of the source
code, as there are no more attach_source/detach_source calls flying
all around in the source code that were called during the
initialization process.