clipboard-rdp: Add public async FileContents request functions
In order to be able to manage the files, that were copied to the
clipboard, via FUSE, add public functions to request the file size of a
file and to request data of a file.
These functions don't handle the FileContentsResponse, which allows
the FUSE implementation for the clipboard to request the data in an
asynchronous way.