session-rdp: Add enum to track status of mouse pointer type
Mouse pointer updates in RDP are not just bitmap updates. The protocol
also allows us to hide the mouse pointer or use the default one that is
always available on the client.

To be able to track the status of the mouse pointer type, add an enum,
which reflects the current state of the pointer.
This allows us in the following commits to only submit an identical
mouse pointer type (e.g. SYSPTR_NULL) once, reducing extra bandwidth
usage.