session-rdp: Add support for unicode input
In addition to keyboard input via keycodes, RDP also supports input
of Unicode characters.
This does not replace keyboard input via keycodes (keyboard input via
keycodes is still the expected way of keyboard input), but allows
clients to input specific characters this way.
Remmina, for example, makes use of it.

To send mutter the Unicode character, convert it first into a keysym
using xkbcommon.

To be able to make use of that conversion function, convert the Unicode
character first from UTF-16 to UTF-32, as xkbcommon can only convert
UTF-32 characters into a keysym.
Then submit the keysym to mutter to input that character.