session-rdp: Send Provider Ultimatum on server sided disconnection
Currently, when the remote desktop connection is stopped via g-s, g-r-d
doesn't notify the client about the reason of the disconnection.
This leaves the remote desktop client in the dark, as it is unable to
handle this situation properly, as it doesn't simply know the cause of
the disconnection.
For example, if the client has auto-reconnection enabled, it will try
to reconnect to g-r-d again on a server sided disconnection even though
the user might have triggered the disconnection by himself via g-s.

Unfortunately, mutter (currently) doesn't provide a reason when the
remote desktop connection was stopped via g-s, so for now always tell
the remote desktop client that the connection was dropped by an
administrator or an administrative tool.