build: Make unicode input in RDP backend optional
Although unicode input is a core functionality for RDP, and is
necessary to be able to type (keyboard input) for some clients
(e.g. Remmina), we need a non-deprecated version of libxkbcommon, to be
able to convert the unicode char into a keysym.
The freedesktop-sdk however, (which is apparently needed for Gnome OS)
can currently only provide an old version of libxkbcommon.

So, make the unicode input optional (breaking functionality) to be able
to make Gnome OS builds.
For Gnome 41, this commit can be reverted again.