rdp: Add event queue
For proper channel support in gnome-remote-desktop, socket events must
be received and handled in a different thread.
This is done in the next commit.
However, this also means that we receive input events in the same new
thread, as checking the file descriptor (and therefore handling its
events) happens in the same function.
Input events are pushed via the dbus handling in the main thread.
Therefore, we have to ensure that these events are handled while the
dbus session proxy still exists.

To ensure that, add an event queue, where the socket thread, that will
be implemented in the next commit, can push all input events and then
queue a flush on the main thread.