clipboard-rdp: Use FUSE instead of downloading all file contents
Currently, all files that are available on the client side are being
transferred to the server when the uri/list for them is being accessed.
This is wasteful in terms of space and bandwidth usage.
It also has the disadvantage that a local application might need to
request the uri-list twice when large files are being transferred,
because the timeout on the mutter side might have been already reached,
while the files are still being transferred.
Another disadvantage is that local applications won't show a progress
bar when transferring the files.

Solve this by replacing the current direct-transfer implementation with
the previously implemented FUSE clipboard.
With FUSE, gnome-remote-desktop advertises the files, that are
available on the client side, in addition to the uri-list, and only
transfers the file content, when the files on the FUSE filesystem are
accessed.