rdp: Add display control class
This class will be used by the Display Control channel, which is a very
simple dynamic channel, as it only has two PDUs.
The first PDU is the DISPLAYCONTROL_CAPS_PDU, which is sent initially
by the server side, containing the constraints of the monitor
configurations.
The second PDU is the DISPLAYCONTROL_MONITOR_LAYOUT_PDU, which is sent
by the client side, to request a monitor layout change.

When receiving a new monitor layout, create a new internal monitor
configuration and use the previous implemented API in session-rdp to
submit the new configuration.