clipboard-rdp: Implement clipboard data locking
Normally, when copying file content via the clipboard in RDP, the file
stream data exists as long as the same file list exists on the other
peers side.
This has the effect that copying new content to the clipboard aborts
the transfer of files that are currently copied via the clipboard to
the other peer.

Clipboard data locking is a mechanism that allows a peer to assign a
clipDataId to the current file list.
This operation is called "locking the clipboard data".
With this "lock" in place, the peer, which received the lock MUST
retain the file stream data (aka the uri-list) for this clipDataId.
When now a FileContentsRequest happens, the requesting peer now also
attaches the clipDataId to the request, which then the other peer looks
up.
When the clipboard content changes, the peer with the lock still knows
now which file list is meant and can therefore still provide file data,
although the new clipboard content might not even have a file list any
more.
Once all operations are done, the peer, which emitted the lock, can
request an unlock to free the clipDataId and the associated file list.

The cliprdr documentation mentions that a lock or an unlock can happen
at any time, as soon as the initialization sequence is complete.
However, the relationship between clipDataId and file list MUST be
bijective.
This means that a peer is not allowed to request multiple locks for the
same file list.
It is also not allowed to try to emit a lock for clipDataId that is
already in use.

While the cliprdr documentation shows the example of using the locking
mechanism before requsting the file list via a FormatDataRequest, it
is not the behaviour that mstsc has.
When locking is advertised, mstsc (, which supports locking,) always
unconditionally emit a lock as soon as it receives a new FormatList.
This lock is emitted before sending the FormatListResponse.
While a clipDataId can only be used in conjunction with a file list,
this behaviour is a little bit overkill, but not forbidden.
As soon as mstsc sent the lock and the FormatListResponse, it will try
to emit an unlock for the previous clipDataId.
If a pending associated file operation is still happening, the unlock
will happen as soon as the operation finishes.

To implement the locking mechanism, gnome-remote-desktop needs to be
able to track these clipDataIds.
Each clipDataId in g-r-d will have a bijective relation to a
ClipDataEntry.
This entry, will track the clipDataId, the associated WinPR clipboard,
since the WinPR clipboard only works with one file list, a boolean
value to track, whether file operations are allowed (FormatListResponse
with CB_RESPONSE_OK flag received) and whether the file list is
available in the WinPR clipboard (Client requested the file list via a
FormatDataRequest before).

ClipDataEntries are managed with two HashTables: one (serial, entry)-
HashTable and one (clipDataId, entry)-HashTable.
The former is used by the main thread to determine the current
ClipDataEntry (if it exists) to allow the entry to request file
contents, to set the state of having a file list or to set the
independence state of the associated WinPR clipboard.
The latter is used by the cliprdr thread for FileContentsRequests.
Upon creating and destroying a ClipDataEntry both HashTables are
synchronized.
This means: If exists (serial, entry), then exists (clipDataId, entry)
too.
And: If exists (clipDataId, entry), then exists (serial, entry) too.

For the usage of the ClipDataEntry no mutex is required, because the
only actions that the main thread does is setting the boolean values
of the independence state (upon sending a new FormatList), the
allowance state (of FileContentsRequests) and the has_file_list state.
This allows the cliprdr thread to handle FileContentsRequests without
depending on the main thread (waiting on a mutex to be released).

Each FormatList, that the server sends, will create a new serial for
the clipboard.
If an entry with that serial exists, the associated WinPR clipboard,
becomes independent from the main one, since the main clipboard will
be recreated.
Upon destroying a ClipDataEntry, the WinPR clipboard will only be
destroyed, when it is independent from the main WinPR clipboard.

While this locking behaviour in theory allows new copy and paste
actions, while a file transfer operation is happening, the reality is
different.
Microsofts locking implementation is poorly done: Locking with mstsc
only works, if the clipboard owner doesn't change.
For this commit specifically, this means that if a server-to-client
transfer is happening and a copy operation on the client side happens,
mstsc will still abort the transfer although g-r-d still has the data
available.
The abort happens on the client side. gnome-remote-desktop cannot
prevent this behaviour.
As soon as FreeRDP implements locking on its side, we can expect to
have a proper implementation, that does not abort the current transfer,
when the clipboard owner changes.