session: Don't submit PointerMotion events when stream is not ready
mstsc submits PointerMotion events very early when it connects to the
server.
In such situations, the stream, which is needed for PointerMotion
events, might not be ready and gnome-remote-desktop will crash due to
dereferencing a NULL pointer.

To fix this, drop PointerMotion events if the stream is not yet ready.