session-rdp: Use the value 10.0 DISCRETE_SCROLL_STEP
Although the value 10.0 is a terrible choice, for the scrolling step,
mutter, unfortunately, uses it everywhere.
Using a different (and usually a better) value like 12 is unfortunately
currently not possible, unless mutter adopts a better scrolling step,
like 120.

So, change the value to 10.0 instead of 12 to not mess things up on the
mutter side, when emulating discrete scrolling events.
Additionally, rename the variable to DISCRETE_SCROLL_STEP, as that is
the name that mutter internally uses.