session-rdp: Add handling for the PAUSE key (XKB_KEY_Pause)
Unlike other key events, the PAUSE key is not submitted as a keycode.
Instead RDP uses a special sequence of the CTRL and Numlock key to
emit the key:

1. CTRL-key DOWN (in combination with the KBD_FLAGS_EXTENDED1 flag)
2. Numlock-key DOWN
3. CTRL-key UP (in combination with the KBD_FLAGS_EXTENDED1 flag)
4. Numlock-key UP

When this sequence is received, emit a PAUSE-key down and a PAUSE-key
release event, as RDP only emits this sequence when the PAUSE-key is
pressed.
When the PAUSE-key is released on the client side, the RDP client won't
notify the server about it, so the PAUSE-key must be released by
gnome-remote-desktop upon sending the key-down event.