clipboard-rdp: Simplify delegate handling
In the first approach of the clipboard implementation for RDP every
function operated on the CliprdrServerContext instance.
This handling got replaced in the final version, but the delegate
handling still seems to use it.

Directly using the delegate instance not only simplifies the delegate
functions, but is also necessary for clipboard locking, as
gnome-remote-desktop might need to use a different delegate than the
default one.