session-rdp: Synchronize kbd lock states when SynchronizeEvent is called
Whenever the RDP client gets focused, the SynchronizeEvent is emitted.
When this event is emitted, the RDP client submits the current states
of its toggle keys.
Additionally, the RDP server resets all keys.
While the latter action is already implemented in gnome-remote-desktop,
the former one is not.

So, implement the former action too by submitting a fake key press and
fake key release of the lock keys, when the current state in mutter
differs from the state that the RDP client submitted.
Do this action using the previously implemented APIs in
GrdRdpEventQueue.