clipboard-rdp: Don't wait for FormatListResponses
Currently, when gnome-remote-desktop sends a new FormatList to the RDP
client, it will wait for the RDP client to send the FormatListResponse
for the previous FormatList, when a response is pending.
However, if the user on the server side decides to hold CTRL+C to spam
new FormatLists and the clipboard handling on the client sides goes up
in flames, gnome-remote-desktop won't get a response any more.
Assuming the number of FormatLists to send is 1000, then
gnome-remote-desktop will block the main thread in total for 4000
seconds, which is obviously an atrocious situation.

Additionally, mutter has three types of clipboards:
The normal one, one for primary selection and one for Drag'n'Drop
usage.
For gnome-remote-desktop only the first one is relevant.
In addition to that for each of these three clipboards there can only
be one clipboard owner.

This characteristic allows gnome-remote-desktop to optimize this
situation:
Whenever, a FormatList is about to be sent, but the previous
FormatListResponse is still pending, queue the new FormatList.
If there is already a FormatList queued, replace it.
To make this feasible, the handling of FormatLists must be revamped to
work in an asynchronous way, as the dbus handling happens in the main
thread and the pending dbus messages are handled inside the dbus
component in gnome-remote-desktop.
Whenever a new FormatList is queued (due to pending
FormatListResponse), create a timeout-source which clears the pending
FormatList upon reaching timeout.
When a FormatListResponse is received, this timeout will be removed.

With this handling, gnome-remote-desktop won't spam further the RDP
client, when it doesn't get a response any more.
It also reduces bandwidth usage and allows the client side to restart
the clipboard handling.
With mstsc, for example, the clipboard handling is done in rdpclip.exe.
The handling of rdpclip.exe is relatively stable and it won't go up in
flames, unless clipboard locking is supported.
With this commit, however, rdpclip.exe won't even do that any more,
since a lot of FormatLists are already sorted out, when FormatLists are
spammed.